requirements.txt
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.0.1
